refactor
- domain + persistence
- move entities to domain
- embedded/inmemory mongodb tests
