--- bff/fs-bff-parent/fs-bff-common/src/main/java/com/makimenko/fs/common/dao/BookDao.java
@@ -1,14 +1,0 @@
-package com.makimenko.fs.common.dao;
-
-
-import com.makimenko.fs.common.domain.book.Book;
-
-import java.util.List;
-
-public interface BookDao {
-
-    List<Book> getAllBooks();
-
-    Book createBook(Book book);
-
-}
