--- bff/fs-bff-parent/fs-bff-common/src/main/java/com/makimenko/fs/common/domain/BaseDomain.java
@@ -1,12 +1,0 @@
-package com.makimenko.fs.common.domain;
-
-import lombok.Data;
-
-import java.io.Serializable;
-
-@Data
-public abstract class BaseDomain<T extends Serializable> implements Serializable {
-
-    private T id;
-
-}
