--- bff/fs-bff-parent/fs-bff-common/src/main/java/com/makimenko/fs/common/domain/BusinessData.java
@@ -1,11 +1,0 @@
-package com.makimenko.fs.common.domain;
-
-import lombok.Data;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-@Data
-public abstract class BusinessData extends BaseDomain<UUID> implements Serializable {
-
-}
