--- bff/fs-bff-parent/fs-bff-common/src/main/java/com/makimenko/fs/common/domain/Person.java
@@ -1,13 +1,0 @@
-package com.makimenko.fs.common.domain;
-
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-public abstract class Person extends BusinessData {
-
-    private String name;
-    private LocalDate birthDate;
-
-}
