--- bff/fs-bff-parent/fs-bff-common/src/main/java/com/makimenko/fs/common/domain/RefData.java
@@ -1,12 +1,0 @@
-package com.makimenko.fs.common.domain;
-
-import lombok.Data;
-
-import java.io.Serializable;
-
-@Data
-public abstract class RefData extends BaseDomain<String> implements Serializable {
-
-    String name;
-
-}
