--- bff/fs-bff-parent/fs-bff-common/src/main/java/com/makimenko/fs/common/domain/book/Genre.java
@@ -1,11 +1,0 @@
-package com.makimenko.fs.common.domain.book;
-
-import com.makimenko.fs.common.domain.RefData;
-import lombok.Data;
-import lombok.ToString;
-
-@Data
-@ToString(callSuper = true)
-public class Genre extends RefData {
-
-}
