+++ bff/fs-bff-parent/fs-bff-domain/src/main/java/com/makimenko/fs/domain/BusinessData.java
@@ -1,0 +1,11 @@
+package com.makimenko.fs.domain;
+
+import lombok.Data;
+
+import java.io.Serializable;
+import java.util.UUID;
+
+@Data
+public abstract class BusinessData extends BaseDomain<UUID> implements Serializable {
+
+}
