+++ bff/fs-bff-parent/fs-bff-domain/src/main/java/com/makimenko/fs/domain/Person.java
@@ -1,0 +1,13 @@
+package com.makimenko.fs.domain;
+
+import lombok.Data;
+
+import java.time.LocalDate;
+
+@Data
+public abstract class Person extends BusinessData {
+
+    private String name;
+    private LocalDate birthDate;
+
+}
