+++ bff/fs-bff-parent/fs-bff-domain/src/main/java/com/makimenko/fs/domain/RefData.java
@@ -1,0 +1,12 @@
+package com.makimenko.fs.domain;
+
+import lombok.Data;
+
+import java.io.Serializable;
+
+@Data
+public abstract class RefData extends BaseDomain<String> implements Serializable {
+
+    String name;
+
+}
