+++ bff/fs-bff-parent/fs-bff-domain/src/main/java/com/makimenko/fs/domain/book/Author.java
@@ -1,0 +1,18 @@
+package com.makimenko.fs.domain.book;
+
+import com.makimenko.fs.domain.Person;
+import lombok.Data;
+import lombok.ToString;
+
+@Data
+@ToString(callSuper = true)
+public class Author extends Person {
+
+    String licenseId;
+    BookGenre bookGenre;
+
+    public static void main(String[] args) {
+
+    }
+
+}
