+++ bff/fs-bff-parent/fs-bff-domain/src/main/java/com/makimenko/fs/domain/book/Book.java
@@ -1,0 +1,19 @@
+package com.makimenko.fs.domain.book;
+
+import com.makimenko.fs.domain.BusinessData;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@NoArgsConstructor
+@Data
+public class Book extends BusinessData {
+
+    private String title;
+    private Author author;
+    private BookGenre bookGenres[];
+    private int avgRank;
+    private LocalDate releaseDate;
+
+}
