+++ bff/fs-bff-parent/fs-bff-domain/src/main/java/com/makimenko/fs/domain/book/BookGenre.java
@@ -1,0 +1,12 @@
+package com.makimenko.fs.domain.book;
+
+
+import com.makimenko.fs.domain.RefData;
+import lombok.Data;
+import lombok.ToString;
+
+@Data
+@ToString(callSuper = true)
+public class BookGenre extends RefData {
+
+}
