--- bff/fs-bff-parent/fs-bff-persistence/fs-bff-persistence-hibernate/src/main/java/com/makimenko/fs/persistence/hibernate/config/MongoDbConfig.java
@@ -1,12 +1,0 @@
-package com.makimenko.fs.persistence.mongodb.config;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan({
-        "com.makimenko.fs.persistence.mongodb"
-})
-public class MongoDbConfig {
-
-}
