--- bff/fs-bff-parent/fs-bff-persistence/fs-bff-persistence-hibernate/src/main/resources/application.yml
@@ -1,14 +1,0 @@
-logging:
-  level:
-    root: INFO
-    com.makimenko.fs: INFO
-
-spring:
-  data:
-    mongodb:
-      host: localhost
-      port: 27017
-      database: fs
-      username: root
-      password: example
-      authentication-database: admin
