--- bff/fs-bff-parent/fs-bff-persistence/fs-bff-persistence-mongodb/src/main/java/com/makimenko/fs/persistence/mongodb/entity/BusinessEntity.java
@@ -1,13 +1,0 @@
-package com.makimenko.fs.persistence.mongodb.entity;
-
-import com.makimenko.fs.common.domain.BusinessData;
-
-import java.io.Serializable;
-
-public abstract class BusinessEntity<T extends BusinessData> implements Serializable {
-
-    public abstract T toDomain();
-
-    public abstract void updateEntity(T dom);
-
-}
