+++ bff/fs-bff-parent/fs-bff-persistence/src/main/java/com/makimenko/fs/persistence/config/MongoDbConfig.java
@@ -1,0 +1,14 @@
+package com.makimenko.fs.persistence.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+
+@Configuration
+@ComponentScan({
+        "com.makimenko.fs.persistence"
+})
+@EnableMongoRepositories({"com.makimenko.fs.persistence.repository"})
+public class MongoDbConfig {
+
+}
