+++ bff/fs-bff-parent/fs-bff-persistence/src/main/java/com/makimenko/fs/persistence/dao/BookDao.java
@@ -1,0 +1,14 @@
+package com.makimenko.fs.persistence.dao;
+
+
+import com.makimenko.fs.common.domain.book.Book;
+
+import java.util.List;
+
+public interface BookDao {
+
+    List<Book> getAllBooks();
+
+    Book createBook(Book book);
+
+}
