+++ bff/fs-bff-parent/fs-bff-persistence/src/test/java/com/makimenko/fs/persistence/TestConfig.java
@@ -1,0 +1,13 @@
+package com.makimenko.fs.persistence;
+
+import com.makimenko.fs.persistence.config.MongoDbConfig;
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.Import;
+
+@SpringBootConfiguration
+@EnableAutoConfiguration
+@Import(MongoDbConfig.class)
+public class TestConfig {
+
+}
