+++ bff/fs-bff-parent/fs-bff-persistence/src/test/resources/application.yml
@@ -1,0 +1,5 @@
+---
+spring:
+  profiles: test
+  data:
+    mongodb:
